+++ tests/test_ddl.py
@@ -76,7 +76,10 @@ def test_modify_column():
     if isinstance(Migrate.ddl, MysqlDDL):
         assert ret0 == "ALTER TABLE `category` MODIFY COLUMN `name` VARCHAR(200)"
     elif isinstance(Migrate.ddl, PostgresDDL):
-        assert ret0 == 'ALTER TABLE "category" ALTER COLUMN "name" TYPE VARCHAR(200)'
+        assert (
+            ret0
+            == 'ALTER TABLE "category" ALTER COLUMN "name" TYPE VARCHAR(200) USING "name"::VARCHAR(200)'
+        )
 
     if isinstance(Migrate.ddl, MysqlDDL):
         assert (
@@ -84,7 +87,9 @@ def test_modify_column():
             == "ALTER TABLE `user` MODIFY COLUMN `is_active` BOOL NOT NULL  COMMENT 'Is Active' DEFAULT 1"
         )
     elif isinstance(Migrate.ddl, PostgresDDL):
-        assert ret1 == 'ALTER TABLE "user" ALTER COLUMN "is_active" TYPE BOOL'
+        assert (
+            ret1 == 'ALTER TABLE "user" ALTER COLUMN "is_active" TYPE BOOL USING "is_active"::BOOL'
+        )
 
 
 def test_alter_column_default():